made the admin panel mobiel responsive

+++ src/common/UI/AdminPage/index.js
@@ -111,30 +111,36 @@ const AdminPanel = () => {
     router.push("/signin");
   };
 
-  const deleteUser = async (userId) => {
-    try {
-      const response = await fetch(apiUrl + "delete_user/", {
-        method: "DELETE",
-        headers: {
-          Authorization: `Bearer ${message}`,
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(userId),
-        cors: "cors",
-      });
-      const data = await response.json();
-      if (response.ok && data.msg === "User successfully deleted") {
-        setRefresh(!refresh);
-        alert("User deleted successfully");
-        // Perform any additional actions needed after successful deletion
-      } else {
-        alert("Failed to delete user");
-        // Handle the error response as needed
-      }
-    } catch (error) {
-      console.error(error);
+const deleteUser = async (userId) => {
+  // Show confirmation prompt to the user
+  const confirmDelete = window.confirm("Are you sure you want to delete this user?");
+  if (!confirmDelete) {
+    return; // If the user cancels the deletion, do not proceed
+  }
+
+  try {
+    const response = await fetch(apiUrl + "delete_user/", {
+      method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${message}`,
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(userId),
+      cors: "cors",
+    });
+    const data = await response.json();
+    if (response.ok && data.msg === "User successfully deleted") {
+      setRefresh(!refresh);
+      alert("User deleted successfully");
+      // Perform any additional actions needed after successful deletion
+    } else {
+      alert("Failed to delete user");
+      // Handle the error response as needed
     }
-  };
+  } catch (error) {
+    console.error(error);
+  }
+};
 
   const manualVerifyUser = async (userId) => {
     try {
@@ -167,50 +173,60 @@ const AdminPanel = () => {
 
   return (
     <div className="bg-gray-100 min-h-screen flex flex-col">
-      <div className="py-4 px-6 flex items-center min-w-full justify-between fixed z-50  bg-white shadow-md">
-        <h1 className="text-3xl font-semibold">COGNOSCO ADMIN PANEL</h1>
-        <div className="flex items-center gap-4">
-          <div
-            className={`px-4 py-2 rounded cursor-pointer ${
-              activeTab === "registered"
-                ? "bg-gray-900 text-white"
-                : "text-gray-900"
-            }`}
-            onClick={() => setActiveTab("registered")}
-          >
-            Registered Users
-          </div>
-          <div
-            className={`px-4 py-2 rounded cursor-pointer ${
-              activeTab === "pending"
-                ? "bg-amber-300 text-white"
-                : "text-gray-900"
-            }`}
-            onClick={() => setActiveTab("pending")}
-          >
-            Pending Verification
-          </div>
-          <div
-            className={`px-4 py-2 rounded cursor-pointer ${
-              activeTab === "Pending Payment"
-                ? "bg-yellow-800 text-white"
-                : "text-gray-900"
-            }`}
-            onClick={() => setActiveTab("Pending Payment")}
-          >
-            Pending Payment
-          </div>
-          <div
-            className="px-4 py-2 rounded cursor-pointer bg-red-700 text-white"
-            onClick={() => {
-              signOut();
-            }}
-          >
-            Logout
-          </div>
-        </div>
-      </div>
-      <div className="flex-grow pt-5  flex flex-col relative top-16 overflow-y-auto">
+<div className="py-4 px-6 flex flex-col  items-center justify-between fixed z-50 bg-white min-w-full shadow-md md:flex-row md:justify-between md:items-center">
+  <h1 className="text-3xl font-semibold">COGNOSCO ADMIN PANEL</h1>
+  <div className="flex flex-col md:flex-row md:items-center gap-1 md:gap-4">
+    <div
+      className={`px-4 py-2 rounded cursor-pointer ${
+        activeTab === "registered"
+          ? "bg-gray-900 text-white"
+          : "text-gray-900"
+      }`}
+      onClick={() => setActiveTab("registered")}
+    >
+      Registered Users
+    </div>
+    <div
+      className={`px-4 py-2 rounded cursor-pointer ${
+        activeTab === "pending"
+          ? "bg-amber-300 text-white"
+          : "text-gray-900"
+      }`}
+      onClick={() => setActiveTab("pending")}
+    >
+      Pending Verification
+    </div>
+    <div
+      className={`px-4 py-2 rounded cursor-pointer ${
+        activeTab === "verified"
+          ? "bg-green-500 text-white"
+          : "text-gray-900"
+      }`}
+      onClick={() => setActiveTab("verified")}
+    >
+      Verified Users
+    </div>
+    <div
+      className={`px-4 py-2 rounded cursor-pointer ${
+        activeTab === "Pending Payment"
+          ? "bg-yellow-800 text-white"
+          : "text-gray-900"
+      }`}
+      onClick={() => setActiveTab("Pending Payment")}
+    >
+      Pending Payment
+    </div>
+    <div
+      className="px-4 py-2 rounded cursor-pointer bg-red-700 text-white"
+      onClick={() => {
+        signOut();
+      }}
+    >
+      Logout
+    </div>
+  </div>
+</div>
+      <div className="flex-grow pt-5  flex flex-col relative md:top-16 top-72 overflow-y-auto">
         {activeTab === "registered" &&
           allUsers.length > 0 &&
           allUsers.map((user) => (
@@ -237,6 +253,43 @@ const AdminPanel = () => {
               </div>
             </div>
           ))}
+        {activeTab === "verified" &&
+          allUsers.length > 0 &&
+          allUsers
+            .filter((user) => user.status === 1) // Filter users with status = 1
+            .map((user) => (
+              <div
+                key={user.id}
+                className="mx-4 my-2 p-4 bg-white shadow-md rounded-lg"
+              >
+                <div className="text-lg font-semibold">{user.name}</div>
+                <div className="text-gray-600">{user.email_id}</div>
+                <div className="text-gray-600">Status: {"Verified"}</div>
+                <div className="text-gray-600">
+                  Phone Number: {user.phone_no}
+                </div>
+                <div className="mt-4 flex gap-4 justify-end">
+                  <button
+                    className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded"
+                    onClick={async () => {
+                      await deleteUser(user.email_id);
+                    }}
+                  >
+                    Delete
+                  </button>
+                   <button
+                  className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
+                  onClick={async () => {
+                    await getUserDetails(user.email_id);
+                    setShowModal(true);
+                  }}
+                >
+                  View UPI Image
+                </button>
+                </div>
+              </div>
+            ))}
+
         {activeTab === "pending" &&
           pendingUsers.length > 0 &&
           pendingUsers.map((user) => (
@@ -278,18 +331,20 @@ const AdminPanel = () => {
               </div>
             </div>
           ))}
-        {activeTab === "pending" &&
-          pendingUsers.length ===0 ? (
-             <div className="mx-4 my-2 p-4 bg-white shadow-md rounded-lg">
-            <p className="text-lg font-semibold text-center">No Pending Verifications. </p>
+        {activeTab === "pending" && pendingUsers.length === 0 ? (
+          <div className="mx-4 my-2 p-4 bg-white shadow-md rounded-lg">
+            <p className="text-lg font-semibold text-center">
+              No Pending Verifications.{" "}
+            </p>
           </div>
-          ): null}
-         {activeTab === "Pending Payment" &&
-          pendingPaymentUsers.length ===0 ? (
-             <div className="mx-4 my-2 p-4 bg-white shadow-md rounded-lg">
-            <p className="text-lg font-semibold text-center">No Pending Payments. </p>
+        ) : null}
+        {activeTab === "Pending Payment" && pendingPaymentUsers.length === 0 ? (
+          <div className="mx-4 my-2 p-4 bg-white shadow-md rounded-lg">
+            <p className="text-lg font-semibold text-center">
+              No Pending Payments.{" "}
+            </p>
           </div>
-          ): null}
+        ) : null}
 
         {activeTab === "Pending Payment" &&
           pendingPaymentUsers.length > 0 &&
